changed to mongodb, implemented goroutines

+++ models/payload.go
@@ -3,7 +3,7 @@ package models
 type CrawlerPayload struct {
 	URL      string   `json:"url"`
 	Keywords []string `json:"keywords"`
-	Depth    int32    `json:"depth"`
+	Depth    int8    `json:"depth"`
 }
 
 type CrawlerRequest struct {
